+++ Library/Compression/Integers/ICodec.cs
@@ -1,6 +1,6 @@
 namespace InvertedTomato.Compression.Integers;
 
-public class VlqCodec : ICodec
+public class ICodec
 {
     
 }
