+++ Library/Compression/Integers/Legacy/FibonacciCodec.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 
-namespace InvertedTomato.Compression.Integers {
+namespace InvertedTomato.Compression.Integers.Legacy {
 	public class FibonacciCodec : Codec {
 		/// <summary>
 		///     Minimum value this codec can support.
