+++ Library/Compression/Integers/Legacy/IByteReader.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace InvertedTomato.Compression.Integers {
+namespace InvertedTomato.Compression.Integers.Legacy {
 	public interface IByteReader {
 		Byte ReadByte();
 	}
