+++ Library/Compression/Integers/Legacy/IByteWriter.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace InvertedTomato.Compression.Integers {
+namespace InvertedTomato.Compression.Integers.Legacy {
 	public interface IByteWriter {
 		void WriteByte(Byte value);
 	}
