+++ Library/Compression/Integers/Legacy/VlqCodec.cs
@@ -1,34 +1,21 @@
-﻿using System;
+using System;
 using System.IO;
 
-namespace InvertedTomato.Compression.Integers {
+namespace InvertedTomato.Compression.Integers.Legacy {
+	
 	/// <summary>
-	/// VLQ similar to https://en.wikipedia.org/wiki/Variable-length_quantity with "Removing Redundancy", but the
-	/// continuation bit flag is reversed. This might be more performant for datasets with consistently large values.
+	/// Traditional VLQ implementation as per https://en.wikipedia.org/wiki/Variable-length_quantity with "Removing Redundancy".
 	/// </summary>
-	public class InvertedVlqCodec : Codec {
-		public const UInt64 MinValue = UInt64.MinValue;
+	public class VlqCodec:Codec {
+		public const UInt64 MinValue = 0;
 		public const UInt64 MaxValue = UInt64.MaxValue - 1;
-		public const Byte Nil = 0x80; // 10000000
 
-		public static readonly Byte[] Zero = new Byte[] {0x80}; // 10000000
-		public static readonly Byte[] One = new Byte[] {0x81}; // 10000001
-		public static readonly Byte[] Two = new Byte[] {0x82}; // 10000010
-		public static readonly Byte[] Four = new Byte[] {0x84}; // 10000100
-		public static readonly Byte[] Eight = new Byte[] {0x88};
-
-		private const Byte Mask = 0x7f; // 01111111
+		private const Byte More = 0b10000000;
+		private const Byte Mask = 0b01111111;
 		private const Int32 PacketSize = 7;
 		private const UInt64 MinPacketValue = UInt64.MaxValue >> (64 - PacketSize);
 
-		/// <summary>
-		/// Encode an array of values and write them to a stream.
-		/// </summary>
-		/// <param name="output">Stream to write encoded values to.</param>
-		/// <param name="values">Values to encode.</param>
-		/// <returns>Total number of bytes values encoded as.</returns>
-		/// <exception cref="ArgumentNullException"></exception>
-		/// <exception cref="OverflowException"></exception>
+		
 		public override void EncodeMany(IByteWriter output, UInt64[] values, Int32 offset, Int32 count) {
 #if DEBUG
 			if (null == output) {
@@ -59,7 +46,7 @@ namespace InvertedTomato.Compression.Integers {
 				// Iterate through input, taking X bits of data each time, aborting when less than X bits left
 				while (value > MinPacketValue) {
 					// Write payload, skipping MSB bit
-					output.WriteByte((Byte) (value & Mask));
+					output.WriteByte((Byte) ((value & Mask) | More));
 
 					// Offset value for next cycle
 					value >>= PacketSize;
@@ -67,22 +54,10 @@ namespace InvertedTomato.Compression.Integers {
 				}
 
 				// Write remaining - marking it as the final byte for symbol
-				output.WriteByte((Byte) (value | Nil));
+				output.WriteByte((Byte) (value & Mask));
 			}
 		}
-
-		/// <summary>
-		/// Decompress a number of VLQ-encoded values from a given stream.
-		/// </summary>
-		/// <param name="input">Stream to decode from.</param>
-		/// <param name="values">Pre-sized array to output decoded values to.</param>
-		/// <param name="offset">Starting position in the output array.</param>
-		/// <param name="count">Number of values to decode.</param>
-		/// <returns></returns>
-		/// <exception cref="ArgumentNullException"></exception>
-		/// <exception cref="ArgumentOutOfRangeException"></exception>
-		/// <exception cref="EndOfStreamException"></exception>
-		/// <exception cref="OverflowException"></exception>
+		
 		public override void DecodeMany(IByteReader input, UInt64[] values, Int32 offset, Int32 count){
 #if DEBUG
 			if (null == input) {
@@ -122,13 +97,13 @@ namespace InvertedTomato.Compression.Integers {
 #if DEBUG
 					// Check for overflow
 					if (symbol < pre) {
-						throw new OverflowException("Input symbol larger than the supported limit of 64 bits. Probable corrupt input.");
+						throw new OverflowException("Symbol is larger than maximum supported value or is corrupt. See UnsignedVlq.MaxValue.");
 					}
 #endif
 
 					// Increment bit offset
 					bit += PacketSize;
-				} while ((b & Nil) == 0); // If not final bit
+				} while ((b & More) > 0); // If not final byte
 
 				// Remove zero offset
 				symbol--;
@@ -137,8 +112,7 @@ namespace InvertedTomato.Compression.Integers {
 				values[i] = symbol;
 			}
 		}
-
-
+		
 		public override Int32? CalculateEncodedBits(UInt64 value) {
 			var packets = (Int32) Math.Ceiling(BitOperation.CountUsed(value) / (Single) PacketSize);
