+++ Library/Compression/Integers/Legacy/Wrappers/ListByteWrapper.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 
-namespace InvertedTomato.Compression.Integers {
+namespace InvertedTomato.Compression.Integers.Legacy {
 	public class ListByteWrapper : IByteWriter {
 		private readonly IList<Byte> Underlying;
