+++ Library/Compression/Integers/Legacy/Wrappers/StreamWrapper.cs
@@ -1,7 +1,7 @@
 using System;
 using System.IO;
 
-namespace InvertedTomato.Compression.Integers {
+namespace InvertedTomato.Compression.Integers.Legacy {
 	public class StreamWrapper : IByteReader, IByteWriter {
 		private readonly Stream Underlying;
