+++ Sample/Program.cs
@@ -1,18 +1,34 @@
 ﻿using InvertedTomato.Compression.Integers;
 
-// Instantiate the codec ready to compress
-Codec codec = new FibonacciCodec(); // Using "InvertedTomato.Compression.Integers.Wave3"
+var vlq = new VlqCodec();
+var td = new ThompsonAlphaCodec();
 
-// Compress data - 3x8 bytes = 24bytes uncompressed
 using var stream = new MemoryStream();
-codec.EncodeMany(stream, new ulong[] {1, 2, 3,});
-Console.WriteLine("Compressed data is " + stream.Length + " bytes"); // Output: Compressed data is 2 bytes
 
-// Decompress data
+// Compress
+using (var writer = new BitWriter(stream))
+{
+    writer.WriteInt32(1, td);
+    writer.WriteInt32(2, td);
+    writer.WriteInt32(3, vlq);
+    writer.WriteInt32(4, td);
+    writer.AlignByte();
+    writer.WriteInt32(2, td);
+}
+
+Console.WriteLine("Compressed data is " + stream.Length + " bytes"); // Output: Compressed data is 2 bytes
 stream.Seek(0, SeekOrigin.Begin);
-var decompressed = new ulong[3];
-codec.DecodeMany(stream, decompressed);
-Console.WriteLine(string.Join(",", decompressed)); // Output: 1,2,3
+
+// Decompress
+using (var reader = new BitReader(stream))
+{
+    reader.ReadInt32(td);
+    reader.ReadInt32(td);
+    reader.ReadInt32(vlq);
+    reader.ReadInt32(td);
+    reader.AlignByte();
+    reader.ReadInt32(td);
+}
 
 Console.WriteLine("Done.");
 Console.ReadKey(true);
